+++ android/src/main/java/com/leeson/image_pickers/activitys/BaseActivity.java
@@ -1,0 +1,184 @@
+package com.leeson.image_pickers.activitys;
+
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.net.Uri;
+import android.os.Build;
+import android.provider.Settings;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+
+
+/**
+ * Created by lisen on 2018/4/12.
+ *
+ * @author lisen < 453354858@qq.com >
+ */
+@SuppressWarnings("all")
+public abstract class BaseActivity extends AppCompatActivity {
+
+    private int REQUEST_CODE_PERMISSION = 0x00001;
+
+
+    /**
+     * 请求权限
+     *
+     * @param permissions 请求的权限
+     * @param requestCode 请求权限的请求码
+     */
+    public void requestPermission(String[] permissions, int requestCode) {
+        this.REQUEST_CODE_PERMISSION = requestCode;
+        if (checkPermissions(permissions)) {
+            permissionSuccess(REQUEST_CODE_PERMISSION);
+        } else {
+            List<String> needPermissions = getDeniedPermissions(permissions);
+            ActivityCompat.requestPermissions(this, needPermissions.toArray(new String[needPermissions.size()]), REQUEST_CODE_PERMISSION);
+        }
+    }
+    /**
+     * 检测所有的权限是否都已授权
+     *
+     * @param permissions
+     * @return
+     */
+    private boolean checkPermissions(String[] permissions) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
+            return true;
+        }
+        for (String permission : permissions) {
+            if (ContextCompat.checkSelfPermission(this, permission) != PackageManager.PERMISSION_GRANTED) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    /**
+     * 获取权限集中需要申请权限的列表
+     *
+     * @param permissions
+     * @return
+     */
+    private List<String> getDeniedPermissions(String[] permissions) {
+        List<String> needRequestPermissionList = new ArrayList<>();
+        for (String permission : permissions) {
+            if (ContextCompat.checkSelfPermission(this, permission) !=
+                    PackageManager.PERMISSION_GRANTED ||
+                    ActivityCompat.shouldShowRequestPermissionRationale(this, permission)) {
+                needRequestPermissionList.add(permission);
+            }
+        }
+        return needRequestPermissionList;
+    }
+
+
+    /**
+     * 系统请求权限回调
+     *
+     * @param requestCode
+     * @param permissions
+     * @param grantResults
+     */
+    @Override
+    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if (requestCode == REQUEST_CODE_PERMISSION) {
+            if (verifyPermissions(grantResults)) {
+                permissionSuccess(REQUEST_CODE_PERMISSION);
+            } else {
+                for (int i = 0; i < permissions.length; i++) {
+                    String permission = permissions[i];
+                    if (!ActivityCompat.shouldShowRequestPermissionRationale(this, permission)){
+//                        当用户设置不在询问，并且勾选拒绝权限后，显示提示对话框
+                        permissonNecessity(REQUEST_CODE_PERMISSION);
+                        return;
+                    }
+                }
+                permissionFail(REQUEST_CODE_PERMISSION);
+            }
+        }
+    }
+
+    /**
+     * 确认所有的权限是否都已授权
+     *
+     * @param grantResults
+     * @return
+     */
+    private boolean verifyPermissions(int[] grantResults) {
+        for (int grantResult : grantResults) {
+            if (grantResult != PackageManager.PERMISSION_GRANTED) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public void showSettingDialog(){
+        showTipsDialog();
+    }
+
+    /**
+     * 当用户设置不在询问，并且勾选拒绝权限后，显示提示对话框
+     */
+    public void showTipsDialog() {
+        new AlertDialog.Builder(this)
+                .setTitle("提示信息")
+                .setCancelable(false)
+                .setMessage("当前应用缺少必要权限，该功能暂时无法使用。如若需要，请单击【确定】按钮前往设置中心进行权限授权。")
+                .setNegativeButton("取消", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.dismiss();
+                        finish();
+                    }
+                })
+                .setPositiveButton("确定", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.dismiss();
+                        startAppSettings();
+                        finish();
+                    }
+                }).show();
+    }
+
+    /**
+     * 启动当前应用设置页面
+     */
+    public void startAppSettings() {
+        Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+        intent.setData(Uri.parse("package:" + getPackageName()));
+        startActivity(intent);
+    }
+
+    /**
+     * 获取权限成功 子类调用
+     *
+     * @param requestCode
+     */
+    public void permissionSuccess(int requestCode) {
+
+    }
+
+    /**
+     * 权限获取失败
+     * @param requestCode
+     */
+    public void permissionFail(int requestCode) {
+    }
+    /**
+     * 必要权限获取失败后(子类页面可以重写，做相应的操作)
+     */
+    public void permissonNecessity(int requestCode){
+
+    }
+
+}