+++ android/src/main/java/com/leeson/image_pickers/provides/MyProvide.java
@@ -1,0 +1,13 @@
+package com.leeson.image_pickers.provides;
+
+
+import androidx.core.content.FileProvider;
+
+/**
+ * Created by lisen on 2018-10-10.
+ *
+ * @author lisen < 453354858@qq.com >
+ */
+
+public class MyProvide extends FileProvider {
+}