+++ android/src/main/java/com/leeson/image_pickers/views/ViewPagerFixed.java
@@ -1,0 +1,42 @@
+package com.leeson.image_pickers.views;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.view.MotionEvent;
+
+import androidx.viewpager.widget.ViewPager;
+
+/**
+ * Created by Administrator on 2017/9/14.
+ * lisen
+ */
+
+public class ViewPagerFixed extends ViewPager {
+    public ViewPagerFixed(Context context) {
+        super(context);
+    }
+
+    public ViewPagerFixed(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    @Override
+    public boolean onTouchEvent(MotionEvent ev) {
+        try {
+            return super.onTouchEvent(ev);
+        } catch (IllegalArgumentException ex) {
+            ex.printStackTrace();
+        }
+        return false;
+    }
+
+    @Override
+    public boolean onInterceptTouchEvent(MotionEvent ev) {
+        try {
+            return super.onInterceptTouchEvent(ev);
+        } catch (IllegalArgumentException ex) {
+            ex.printStackTrace();
+        }
+        return false;
+    }
+}