+++ lib/CorpConfig.dart
@@ -1,0 +1,18 @@
+
+///裁剪配置
+///Crop configuration
+class CorpConfig {
+  ///是否可裁剪
+  bool enableCrop = false;
+
+  ///裁剪的宽度比例
+  ///Cropped width ratio
+  int width = 1;
+
+  ///裁剪的高度比例
+  ///Crop height ratio
+  int height = 1;
+
+  CorpConfig({this.enableCrop : false, this.width : 1, this.height : 1});
+
+}