+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 1.0.2
+* Support Android iOS
+* Android i OS supports local language environment (Chinese or English)
+* image_pickers Support picture selection, video multiple selection, support to save network pictures to albums, support preview video and preview picture function